+++ Changes.md
@@ -4,4 +4,4 @@
 Version History
 -------------------
 * v0.1 First working version
-  1. Support Windows and Linux platform.
+  1. Support Windows, Linux and macOS platform.